+++ src/browser/Session.zig
@@ -134,7 +134,7 @@ pub fn createPage(self: *Session) !*Frame {
 
 pub fn removePage(self: *Session) void {
     lp.assert(self.page != null, "Session.removePage - page is null", .{});
-    if (self.page.?.frame._script_manager.is_evaluating) {
+    if (self.page.?.frame._script_manager.base.is_evaluating) {
         // Reentrant teardown from a CDP message drained inside syncRequest;
         // Session.deinit reclaims the page when the connection closes.
         return;
