+++ src/browser/tests/worker/import-module.js
@@ -1,0 +1,4 @@
+export const message = 'imported from module';
+export function multiply(a, b) {
+  return a * b;
+}
