+++ src/browser/tests/worker/modules/base.js
@@ -1,0 +1,1 @@
+export const baseValue = 'from-base';
