+++ src/browser/tests/worker/modules/circular-a.js
@@ -1,0 +1,7 @@
+import { getBValue } from './circular-b.js';
+
+export const aValue = 'a';
+
+export function getFromB() {
+  return getBValue();
+}
