+++ src/browser/tests/worker/modules/circular-b.js
@@ -1,0 +1,11 @@
+import { aValue } from './circular-a.js';
+
+export const bValue = 'b';
+
+export function getBValue() {
+  return bValue;
+}
+
+export function getFromA() {
+  return aValue;
+}
