+++ src/browser/tests/worker/modules/importer.js
@@ -1,0 +1,4 @@
+import { baseValue } from './base.js';
+
+export const importedValue = baseValue;
+export const localValue = 'local';
