+++ src/browser/tests/worker/modules/meta.js
@@ -1,0 +1,1 @@
+export const moduleUrl = import.meta.url;
