+++ src/browser/tests/worker/modules/re-exporter.js
@@ -1,0 +1,2 @@
+export { baseValue } from './base.js';
+export { importedValue, localValue } from './importer.js';
