+++ src/browser/tests/worker/modules/shared.js
@@ -1,0 +1,9 @@
+let counter = 0;
+
+export function increment() {
+  return ++counter;
+}
+
+export function getCount() {
+  return counter;
+}
